Fyp 39 access token

+++ controllers/vehicle_images.js
@@ -2,6 +2,11 @@ const { VehicleImages } = require('../models');
 
 // Create new vehicle images
 const createVehicleImage = async (req, res) => {
+  const { role } = req.user;
+
+  if (role != 'tenant') {
+    return handleErrorResponse(res, 401, 'Token is not valid for Tenant');
+  }
   try {
     const vehicleImages = await VehicleImages.create(req.body);
     return res.status(201).json(vehicleImages);
@@ -37,6 +42,11 @@ const getVehicleImageById = async (req, res) => {
 // Update vehicle images by ID
 const updateVehicleImage = async (req, res) => {
   const vehicleImagesId = req.params.id;
+  const { role } = req.user;
+
+  if (role != 'tenant') {
+    return handleErrorResponse(res, 401, 'Token is not valid for Tenant');
+  }
   try {
     const [updatedRows] = await VehicleImages.update(req.body, {
       where: { id: vehicleImagesId },
@@ -54,6 +64,11 @@ const updateVehicleImage = async (req, res) => {
 // Delete vehicle images by ID
 const deleteVehicleImageById = async (req, res) => {
   const vehicleImagesId = req.params.id;
+  const { role } = req.user;
+
+  if (role != 'tenant') {
+    return handleErrorResponse(res, 401, 'Token is not valid for Tenant');
+  }
   try {
     const deletedRows = await VehicleImages.destroy({
       where: { id: vehicleImagesId },
